commands.go: extract commands to their own files
Commands.go was very long making it difficult to both read and reason
about. Thus, extract each of the main commands into their own file.
